--- src/MicrosoftTeams.interface.ts
@@ -1,9 +1,0 @@
-export const enum NotificationTypes {
-  fileDownloadStart = "fileDownloadStart",
-  fileDownloadComplete = "fileDownloadComplete"
-}
-
-export interface ShowNotificationParameters {
-  message: string;
-  notificationType: NotificationTypes;
-}
